requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# PDF Processing
pypdf2==3.0.1
pdfplumber==0.10.3

# Vector Store & Embeddings
chromadb==0.4.18
openai>=1.6.1
tiktoken==0.5.2

# Optional: Langchain (for future enhancements)
# langchain==0.1.0
# langchain-openai==0.0.2
# langchain-community==0.0.10

# Text Processing (optional)
# nltk==3.8.1
# sentence-transformers==2.2.2

# Utilities
python-dotenv==1.0.0
numpy==1.26.2
pandas==2.1.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# HTTP Requests (for health checks)
requests==2.31.0
